File from Jupyter notebook
From the dataset of ipl matches 2008 to 2020, observe the contents inside the table and finding the duplicate values, unique values, null values in the table. Removing the empty values in the table and removing it. To understand easily visualize the graphs for various data like "Man of the match for most times", "Number of times win by runs", Number of times win by wickets" by using various libraries like numpy, pandas, seaborn, plotly express. Finally, organise all the data need for dashboard in a single table and create the dashboard using Streamlit library in pycharm.